dioxus_hooks

Trait Dependency

pub trait Dependency: Sized + Clone {
    type Out: Clone + PartialEq + 'static;

    // Required method
    fn out(&self) -> Self::Out;

    // Provided method
    fn changed(&self, other: &Self::Out) -> bool { ... }
}

A dependency is a trait that can be used to determine if a effect or selector should be re-run.

Required Associated Types

type Out: Clone + PartialEq + 'static

The output of the dependency

Required Methods

fn out(&self) -> Self::Out

Returns the output of the dependency.

Provided Methods

fn changed(&self, other: &Self::Out) -> bool

Returns true if the dependency has changed.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.