dioxus_hooks

Trait DependencyElement

pub trait DependencyElement:
    'static
    + PartialEq
    + Clone { }

A dependency is a trait that can be used to determine if a effect or selector should be re-run.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.